// src/adapter/interface/schema.interface.ts
export interface IPolicyAdapterSchema {
    name: string
    policy: string
    citizen: string
    license: string
    province: string
    create: Date
    address: string
    brand: string
    model: string
    chassis: string
    doc: string
    start: string
    expire: string
    birth: string
    age: string
    career: string
    email: string
    mobile: string
    seat: string
    weight: string
    disp: string
    transaction: string
}